@@ -9,7 +9,7 @@ using System.Threading.Tasks;
 
 namespace ColoryrServer.DataBase
 {
-    internal class MSCon
+    public class MSCon
     {
         /// <summary>
         /// 连接状态
