@@ -8,7 +8,7 @@ namespace ColoryrServer.DllManager.StartGen.GenUtils
     /// <summary>
     /// App资源存储
     /// </summary>
-    internal record AppBuildSave
+    public record AppBuildSave
     {
         public byte[] Dll { get; set; }
         public byte[] Pdb { get; set; }
@@ -16,7 +16,7 @@ namespace ColoryrServer.DllManager.StartGen.GenUtils
     /// <summary>
     /// 编译后存储
     /// </summary>
-    internal record DllBuildSave
+    public record DllBuildSave
     {
         public AssemblyLoadContext Assembly { get; set; }
         public Type DllType { get; set; }
